app/requirements.txt
python-dotenv==0.15.0
cohere==4.16.0

langchain==0.2.1
langchain-community==0.2.1
transformers==4.31.0
tiktoken==0.3.1

faiss-cpu==1.7.4
llama-index==0.7.10.post1

pandas==1.5.3
scipy==1.10.1
matplotlib==3.7.1
plotly==5.9.0
rich==13.4.2
scikit-learn==1.2.2

mysql-connector-python==8.0.20


pypdf==3.12.2
unstructured==0.8.1
tabulate==0.9.0

Pillow==9.4.0
streamlit==1.24.0
streamlit-chat==0.1.1
streamlit-pills==0.3.0
streamlit_option_menu==0.3.6

altair==4.0

beautifulsoup4==4.12.2
pip